wap-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator>=2.0.0
python-dotenv==1.0.0
azure-communication-email==1.0.0

# ── Azure AI Services ──────────────────────────────────────
openai==1.12.0
azure-search-documents==11.4.0
azure-identity==1.15.0

# ── Azure — new services (Phase 3 migration) ──────────────
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
python-jose[cryptography]==3.3.0
azure-keyvault-secrets==4.7.0
opencensus-ext-azure==1.1.13

# ── Redis ──────────────────────────────────────────────────
redis==5.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Linting and formatting
ruff==0.1.6
black==23.11.0
mypy==1.7.1

